Refactor changes into container_utils
